Move View class to view.py
