+++ nagra/__init__.py
@@ -1,5 +1,6 @@
 from .statement import Statement
-from .table import Table, View
+from .table import Table
+from .view import View
 from .schema import Schema
 from .transaction import Transaction
